Moved Globalnet connectivity validation to framework
