+++ test/e2e/framework/dataplane.go
@@ -15,7 +15,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package dataplane
+package framework
 
 import (
 	"fmt"
